Add quantities to gtins
To support multipacks, add columns add_quantity and remove_quantity to
gtins table. For a 12-pack box of soda cans, for example, the barcode on
the box will equate to 12 instances of the product on either add or
remove, while the barcode on each can will correspond to 1 instance
while adding or removing.

In some cases, there is no outer barcode (no box or shrinkwrap), but the
barcode on the individual item is scanned to represent the entire
multipack. For an 8 pack of a sports drink, we'll count a single scan of
the bottle barcode as 8 instances of the item when we add it to storage,
but only 1 as we remove it, hence two separate columns.

Default is 1.