Add support for item tags and barcode ranges
The database schema already includes support for tags and for barcode ranges, so expose those in the web interface.

"tags" is for adding quick attributes to items, like "quick meals" or "vegetarian" -- we'll expose those for search later.

"barcode ranges" -- certain items, like meats sold by weight, start with a fixed prefix and encode a price in the rest of the UPC. By setting lower and upper values, any barcode scanned in that range, which does not match an exact
code in gtins, will be considered an instance of the item linked to the range.